feat(next-class) styled next class section, responsive on diff screens
