+++ src/App.css
@@ -3,18 +3,14 @@
 }
 
 .logo {
-  width: 100px;
-}
-
-.m-12 {
-  margin: 12px;
+  max-width: 100px;
 }
 
 .App-header {
   font-family: "Nova Square", display;
   display: flex;
-  align-items: center;
-  padding: 0 60px;
+  align-items: start;
+  padding: 40px 60px 0 60px;
   justify-content: space-between;
   font-size: calc(10px + 2vmin);
   background-color: #222222;
@@ -26,20 +22,33 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  font-size: 20px;
+  box-shadow: 0 0 6px #abb1ac40;
+  background-color: #1d1d1d;
+  padding: 0 30px 12px 30px;
 }
 
-@media only screen and (min-width: 960px) {
+@media only screen and (max-width: 700px) {
   .App-header {
-    padding-top: 20px;
+    justify-content: center;
+    flex-direction: column-reverse;
+    align-items: center;
+    padding: 0;
   }
-}
 
-@media only screen and (max-width: 620px) {
-  .logo-container {
+  .box {
+    flex: 0 0 100%;
+  }
+
+  .box:nth-child(2) {
+    order: 0; /* Moves current class to top */
+  }
+
+  .logo {
+    max-width: 60px;
+  }
+
+  .clock-wrapper {
     display: none;
-  }
-
-  .App-header {
-   justify-content: center;
   }
 }
