+++ src/components/CurrentClass.css
@@ -1,0 +1,7 @@
+.title {
+  margin: 0 0 18px 0;
+}
+
+.current-class-container p {
+  margin: 12px;
+}
