+++ src/components/NextClass.css
@@ -1,5 +1,32 @@
-@media only screen and (max-width: 960px) {
+.next-class {
+  font-size: 20px;
+  box-shadow: 0 0 6px #abb1ac40;
+  background-color: #1d1d1d;
+  padding: 6px 30px;
+}
+
+.clock-container {
+  display: none;
+}
+
+.next-class-container {
+  display: flex;
+  justify-content: space-evenly;
+}
+
+@media only screen and (max-width: 700px) {
+  .next-class {
+    font-size: 12px;
+    margin-bottom: 20px;
+    max-width: 120px;
+  }
+
+  .clock-container {
+    display: inline;
+  }
+
   .next-class-container {
-    display: none;
+    width: 100%;
+    margin-bottom: 20px;
   }
 }
