+++ src/components/UsersTable.css
@@ -1,11 +1,11 @@
 .wrapper {
   display: flex;
-  width: 100%;
+  width: 90%;
   flex-wrap: wrap;
   justify-content: center;
   align-content: center;
   gap: 20px 30px;
-  margin-bottom: 60px;
+  margin: 0 auto;
 }
 
 .users-table {
@@ -33,7 +33,7 @@
 
 @media only screen and (max-width: 600px) {
   .users-table {
-    width: 70%;
+    width: 80%;
   }
 }
